requirements.txt
Django==4.2.11
Markdown==3.6
djangorestframework==3.15.2
pillow==10.3.0
django-widget-tweaks==1.5.0
django-mdeditor==0.1.20